Fix bitsandbytes version and deprecated torch_dtype kwarg
All experiments crashed on Kaggle T4 due to:
1. bitsandbytes too old for 4-bit quantization (need >=0.46.1)
2. torch_dtype renamed to dtype in transformers>=4.46
Also fixes KeyError in agent.py when results.tsv has malformed rows.

+++ train.py
@@ -124,7 +124,7 @@ def _run_training() -> None:
         model_id,
         quantization_config=bnb_config,
         device_map="auto",
-        torch_dtype=compute_dtype,
+        dtype=compute_dtype,
     )
 
     # -- Vanilla mode: skip LoRA, eval-only (no training) ----------------------
